M    doit
If a W3P repository directory doesn't exist, git clone it... Before it assumed that if one was missing,
they all were.  Now just check and update one at a time. This probably only occurs if an error happened
during the initial checkout, or if a new lib is added to the 3P.



git-svn-id: https://gforge.sci.utah.edu/svn/uintah/trunk@56343 aee3fe44-4ef4-0310-ac56-cc817e9d0e43