Both temporal and spatial scheduling working correctly for Radiometer.
Effectively, spatial tasks are scheduled "level->eachSpatialPatch()" as opposed to "level->eachPatch". Working to generalize this within the infratructure. For now, schedule spatial tasks independently (independent task for each patch the spatial task runs on), as opposed to scheduling with a spatial PatchSubset. This is so the load balancer can distribute patches more effectively (as opposed to lumping all spatial tasks on a particular node), and so that MPI dependencies can be generated and met correctly.

git-svn-id: https://gforge.sci.utah.edu/svn/uintah/trunk@56764 aee3fe44-4ef4-0310-ac56-cc817e9d0e43