* Updated KokkosOpenMPScheduler.cc to use Kokkos::OpenMP::partition_master() for Kokkos-based builds. * Added -npartitions and -nthreadsperpartition run-time options. * Set the Kokkos-OpenMP Scheduler as the default scheduler for Kokkos-based builds. * Added bulletproofing within SchedulerFactory.cc. * Updated ProcessorGroup.h to use m_comm for -nthreads 1 or -npartitions 1. * Added MasterLock.h to help avoid mixing synchronization primitives. * Replaced use of #include <mutex> with #include <Core/Parallel/MasterLock.h>. * Replaced use of std::mutex with Uintah::MasterLock. * Cleaned up whitespace.
This commit extends the Kokkos-OpenMP Scheduler to support parallel execution of tasks within an MPI process.

Scheduler logic is based upon that of the Unified Scheduler, where individual threads are used as task runners.

In the context of the Kokkos-OpenMP Scheduler, individual partitions are used as task runners.

Individual partitions execute Kokkos-based tasks in parallel across the threads belonging to that partition and existing serial tasks serially.

Note, this scheduler requires MPI_THREAD_MULTIPLE support to run with multiple partitions per MPI process.

git-svn-id: https://gforge.sci.utah.edu/svn/uintah/trunk@57731 aee3fe44-4ef4-0310-ac56-cc817e9d0e43